Deduplicate tabs code
