+++ src/Parser/EnhancementShaman/Modules/Features/ProcTracker.js
@@ -1,6 +1,11 @@
+import React from 'react';
+
 import SPELLS from 'common/SPELLS';
 
 import CoreCooldownTracker, { BUILT_IN_SUMMARY_TYPES } from 'Parser/Core/Modules/CooldownTracker';
+
+import Tab from 'Main/Tab';
+import CooldownOverview from 'Main/CooldownOverview';
 
 class ProcTracker extends CoreCooldownTracker {
   static cooldownSpells = [
@@ -23,6 +28,23 @@ class ProcTracker extends CoreCooldownTracker {
       ],
     },
   ];
+
+  tab() {
+    return {
+      title: 'Procs',
+      url: 'procs',
+      render: () => (
+        <Tab title="Cooldowns">
+          <CooldownOverview
+            fightStart={this.owner.fight.start_time}
+            fightEnd={this.owner.fight.end_time}
+            cooldowns={this.pastCooldowns}
+            showOutputStatistics
+          />
+        </Tab>
+      ),
+    };
+  }
 }
 
 export default ProcTracker;
