--- src/Parser/RetributionPaladin/CONFIG.js
@@ -1,11 +1,0 @@
-import SPECS from 'common/SPECS';
-
-import CombatLogParser from './CombatLogParser';
-import TALENT_DESCRIPTIONS from './TALENT_DESCRIPTIONS';
-
-export default {
-  spec: SPECS.RETRIBUTION_PALADIN,
-  maintainer: '@Zerotorescue',
-  parser: CombatLogParser,
-  talentDescriptions: TALENT_DESCRIPTIONS,
-};
